CharOperatorExample

+++ chap03/src/sec04/exam01_arithmetic/ArtihemticOperatorExample.java
@@ -21,7 +21,7 @@ public class ArtihemticOperatorExample {
 		int result5 = v1 % v2;
 		System.out.println("result5=" + result5);
 		
-		int result6 = (double) v1 / v2;
+		double result6 = (double) v1 / v2;
 		System.out.println("result6=" + result6);
